Refactor cli internals: shared package-command helpers + unified walker
Addresses review findings:
- Walker logic was duplicated between app-data.ts (used by share/html)
  and web-shell.ts (used by web/desktop). Extracted into a single
  project-files.ts; web-shell.ts shrinks to just the index.html
  template.
- web.ts and desktop.ts were ~60% duplicated (arg schema, input
  validation, requirements parsing, destDir reset, try/catch shape).
  New package-command.ts exposes `packageCommandBuilder(yargs)` for
  composing the shared option set, `runPackageCommand(argv, name, body)`
  for the boilerplate handler, and `resolvePackageBuildDir` /
  `collectWheelPaths` / `copyTreeFiltered` helpers. Each command's
  handler now expresses just its host-specific body.
- `createRequire` setup is consistent — declared at module top in
  every file that needs it (no longer relying on the esbuild banner
  in some places and explicit setup in others).
- Missing upstream build dirs (`@stlite/desktop/build`, its `wheels/`,
  `@stlite/browser/build`) now surface a friendly "run `make desktop`"
  error instead of an opaque ENOENT.
- web.ts now skips @stlite/browser's `demos/` and `dev-files/`
  subdirectories when copying the SPA into the artifact, keeping the
  packaged output lean.

CLI surface is unchanged.
