+++ packages/cli/src/web-shell.ts
@@ -1,32 +1,6 @@
 import * as fs from "node:fs";
 import * as path from "node:path";
 
-const IGNORED_DIRS = new Set([".git", "__pycache__", ".venv", "node_modules"]);
-const isIgnoredFile = (name: string) => name.endsWith(".pyc");
-
-/**
- * Walks a project directory and returns POSIX-style paths of every file,
- * filtering out the same defaults as `share`/`html`'s walker.
- */
-export function collectProjectFiles(projectDir: string): string[] {
-  const out: string[] = [];
-  walk(projectDir, [], out);
-  return out;
-}
-
-function walk(rootDir: string, relParts: string[], out: string[]): void {
-  const dirAbs = path.join(rootDir, ...relParts);
-  for (const entry of fs.readdirSync(dirAbs, { withFileTypes: true })) {
-    if (entry.isDirectory()) {
-      if (IGNORED_DIRS.has(entry.name)) continue;
-      walk(rootDir, [...relParts, entry.name], out);
-    } else if (entry.isFile()) {
-      if (isIgnoredFile(entry.name)) continue;
-      out.push([...relParts, entry.name].join("/"));
-    }
-  }
-}
-
 export interface WriteWebIndexHtmlOptions {
   /** Output directory the index.html lives in. */
   destDir: string;
